+++ src/Twilio/Rest/Api/V2010/Account/Call/NotificationInstance.php
@@ -108,16 +108,6 @@ class NotificationInstance extends InstanceResource {
     }
 
     /**
-     * Deletes the NotificationInstance
-     *
-     * @return boolean True if delete succeeds, false otherwise
-     * @throws TwilioException When an HTTP error occurs.
-     */
-    public function delete() {
-        return $this->proxy()->delete();
-    }
-
-    /**
      * Magic getter to access properties
      *
      * @param string $name Property to access