+++ src/Twilio/Rest/Preview/BulkExports/Export/DayInstance.php
@@ -28,7 +28,7 @@ class DayInstance extends InstanceResource {
      *
      * @param \Twilio\Version $version Version that contains the resource
      * @param mixed[] $payload The response payload
-     * @param string $resourceType The resource_type
+     * @param string $resourceType The type of communication – Messages, Calls
      * @return \Twilio\Rest\Preview\BulkExports\Export\DayInstance
      */
     public function __construct(Version $version, array $payload, $resourceType) {