+++ src/Twilio/Rest/Verify/V2/Service/VerificationList.php
@@ -57,6 +57,7 @@ class VerificationList extends ListResource {
             'Payee' => $options['payee'],
             'RateLimits' => Serialize::jsonObject($options['rateLimits']),
             'ChannelConfiguration' => Serialize::jsonObject($options['channelConfiguration']),
+            'AppHash' => $options['appHash'],
         ));
 
         $payload = $this->version->create(