+++ src/Twilio/Rest/Verify/V2/ServiceInstance.php
@@ -26,6 +26,7 @@ use Twilio\Version;
  * @property bool $skipSmsToLandlines
  * @property bool $dtmfInputRequired
  * @property string $ttsName
+ * @property bool $doNotShareWarningEnabled
  * @property \DateTime $dateCreated
  * @property \DateTime $dateUpdated
  * @property string $url
@@ -59,6 +60,7 @@ class ServiceInstance extends InstanceResource {
             'skipSmsToLandlines' => Values::array_get($payload, 'skip_sms_to_landlines'),
             'dtmfInputRequired' => Values::array_get($payload, 'dtmf_input_required'),
             'ttsName' => Values::array_get($payload, 'tts_name'),
+            'doNotShareWarningEnabled' => Values::array_get($payload, 'do_not_share_warning_enabled'),
             'dateCreated' => Deserialize::dateTime(Values::array_get($payload, 'date_created')),
             'dateUpdated' => Deserialize::dateTime(Values::array_get($payload, 'date_updated')),
             'url' => Values::array_get($payload, 'url'),