+++ src/Twilio/Rest/Verify/V2/ServiceList.php
@@ -51,6 +51,7 @@ class ServiceList extends ListResource {
             'DtmfInputRequired' => Serialize::booleanToString($options['dtmfInputRequired']),
             'TtsName' => $options['ttsName'],
             'Psd2Enabled' => Serialize::booleanToString($options['psd2Enabled']),
+            'DoNotShareWarningEnabled' => Serialize::booleanToString($options['doNotShareWarningEnabled']),
         ));
 
         $payload = $this->version->create(