+++ tests/Twilio/Integration/Preview/Bulkexports/Export/JobTest.php
@@ -42,6 +42,9 @@ class JobTest extends HolodeckTestCase {
                 "friendly_name": "friendly_name",
                 "end_day": "end_day",
                 "details": {},
+                "webhook_url": "webhook_url",
+                "webhook_method": "webhook_method",
+                "email": "email",
                 "resource_type": "resource_type"
             }
             '